enh: bands working without plotly installed
