+++ sisl/viz/drawers/_plot_drawers.py
@@ -1,4 +1,4 @@
-from abc import ABC
+from abc import ABC, abstractmethod
 
 from ..configurable import _populate_with_settings
 
@@ -8,25 +8,26 @@ class Drawers:
 
     def __init__(self, plot_cls):
         self._drawers = {}
-        self._drawer = None
 
         self._plot_cls = plot_cls
         self._plot_cls_params = self._plot_cls._get_class_params()
         self._plot_cls_params_keys = [param["key"] for param in self._plot_cls_params[0]]
 
+        self._plot_cls._drawer = None
+
     def register(self, drawer_name, drawer):
+        drawer._drawer_name = drawer_name
         self._drawers[drawer_name] = drawer
 
     def setup(self, plot, drawer_name):
-        self._drawer = self._drawers[drawer_name]()
-
-    def __getattr__(self, key):
-        return getattr(self._drawer, key)
+        current_drawer = getattr(plot, "_drawer", None)
+        if current_drawer is None or current_drawer._drawer_name != drawer_name:
+            if drawer_name not in self._drawers:
+                raise NotImplementedError(f"There is no '{drawer_name}' drawer implemented for {self._plot_cls.__name__} or the drawer has not been loaded.")
+            plot._drawer = self._drawers[drawer_name]()
 
 class Drawer(ABC):
 
-    def setup(self):
-        """Initializes all the things that the drawer needs"""
-
+    @abstractmethod
     def clear(self):
         """Clears the figure so that we can draw again."""
