+++ sisl/viz/plot.py
@@ -17,7 +17,6 @@ from functools import partial
 from pathlib import Path
 
 import dill
-from plotly.subplots import make_subplots
 
 import sisl
 from sisl.messages import info, warn
@@ -83,9 +82,6 @@ class PlotMeta(ConfigurableMeta):
                     else:
                         raise NotImplementedError(
                             f'There is no plot implementation for {sile.__class__} yet.')
-
-            elif isinstance(args[0], go.Figure):
-                plot = Plot.from_plotly(args[0]).update_settings(**kwargs)
             elif isinstance(args[0], Plot):
                 plot = args[0].update_settings(**kwargs)
             else:
@@ -642,9 +638,6 @@ class Plot(ShortCutable, Configurable, metaclass=PlotMeta):
         # Inform whether the plot is in debug mode or not:
         self._debug = _debug
 
-        # Initialize the drawer
-        self._drawers.setup(self, self.get_setting("drawer"))
-
         # Initialize shortcut management
         ShortCutable.__init__(self)
 
@@ -730,34 +723,31 @@ class Plot(ShortCutable, Configurable, metaclass=PlotMeta):
         """ This method is executed only after python has found that there is no such attribute in the instance
 
         So let's try to find it elsewhere. There are two options:
-            - The attribute is in the figure object (self.figure)
+            - The attribute is in the drawer object (self._drawer)
             - The attribute is currently being shared with other plots (only possible if it's a childplot)
         """
-        if key in ["_drawers", "shared_attr"]:
+        if key in ["_drawer", "_get_shared_attr"]:
             pass
-        elif hasattr(self._drawers, key):
-            return getattr(self._drawers, key)
+        elif hasattr(self._drawer, key):
+            return getattr(self._drawer, key)
         else:
             #If it is a childPlot, maybe the attribute is in the shared storage to save memory and time
             try:
-                return self.shared_attr(key)
+                return self._get_shared_attr(key)
             except (KeyError, AttributeError):
                 pass
 
-        raise AttributeError(f"The attribute '{key}' was not found either in the plot, its figure, or in shared attributes.")
+        raise AttributeError(f"The attribute '{key}' was not found either in the plot, its drawer, or in shared attributes.")
 
     def __setattr__(self, key, val):
-        """ If the attribute is one of ["data", "layout", "frames"] we are going to store it directly in `self.figure` for convenience and in order to save memory.
-
+        """
         If is a childplot and it has the attribute `_SHOULD_SHARE_WITH_SIBLINGS` set to True, we will submit the attribute to the shared store.
         This happens in animations/multiple plots. There's a "leading plot" that reads the data and then shares it with the rest
         so that they don't need to read it again, in a collective effort to save memory and time.
 
         Otherwise we set the attribute to the plot itself.
         """
-        if key in ["data", "layout", "frames"]:
-            self.figure.update(**{key: val})
-        elif key != '_SHOULD_SHARE_WITH_SIBLINGS' and getattr(self, '_SHOULD_SHARE_WITH_SIBLINGS', False):
+        if key != '_SHOULD_SHARE_WITH_SIBLINGS' and getattr(self, '_SHOULD_SHARE_WITH_SIBLINGS', False):
             self.share_attr(key, val)
         else:
             object.__setattr__(self, key, val)
@@ -991,8 +981,7 @@ class Plot(ShortCutable, Configurable, metaclass=PlotMeta):
 
         if show and fig_widget is None:
             fig = pt.show(return_figWidget=True)
-            if isinstance(fig, go.FigureWidget):
-                fig_widget = fig
+            fig_widget = fig
 
         if notify:
             trigger_notification("SISL", "Notifications will appear here")
@@ -1017,7 +1006,7 @@ class Plot(ShortCutable, Configurable, metaclass=PlotMeta):
                         if show and fig_widget is None:
                             pt.show()
                         else:
-                            pt._update_FigureWidget(fig_widget)
+                            pt._drawer._update_ipywidget(fig_widget)
 
                         if not forever:
                             self._listening_task.cancel()
@@ -1125,10 +1114,8 @@ class Plot(ShortCutable, Configurable, metaclass=PlotMeta):
 
         If everything is succesful, it calls the next step in plotting (`get_figure`)
         """
-        # Clear all the traces from the figure before drawing the new ones
-        self.clear()
 
-        self._set_data()
+        self._for_drawer = self._set_data()
 
         if update_fig:
             self.get_figure()
@@ -1136,7 +1123,7 @@ class Plot(ShortCutable, Configurable, metaclass=PlotMeta):
         return self
 
     @vizplotly_settings('before')
-    def get_figure(self, **kwargs):
+    def get_figure(self, drawer, **kwargs):
         """
         This is here mainly for historic reasons. At the beggining, the plot's figure was not
         constructed until this point (each time a new figure was constructed). However, it seemed
@@ -1155,6 +1142,14 @@ class Plot(ShortCutable, Configurable, metaclass=PlotMeta):
         self.figure: plotly.graph_objs.Figure
             the plotly figure.
         """
+        # Initialize the drawer
+        self._drawers.setup(self, drawer)
+
+        # Clear all the traces from the figure before drawing the new ones
+        self.clear()
+
+        self.draw(getattr(self, "_for_drawer", None))
+
         call_method_if_present(self, '_after_get_figure')
 
         call_method_if_present(self, 'on_figure_change')
@@ -1186,7 +1181,7 @@ class Plot(ShortCutable, Configurable, metaclass=PlotMeta):
             except Exception as e:
                 warn(e)
 
-        return self._drawers.show(*args, **kwargs)
+        return self._drawer.show(*args, **kwargs)
 
     def _ipython_display_(self, return_figWidget=False, **kwargs):
         """ Handles all things needed to display the plot in a jupyter notebook
@@ -1204,18 +1199,17 @@ class Plot(ShortCutable, Configurable, metaclass=PlotMeta):
 
         def _try_drawer():
             kwargs.pop("listen", None)
-            display_method = getattr(self._drawers, "_ipython_display_", None)
+            display_method = getattr(self._drawer, "_ipython_display_", None)
             if display_method is not None:
                 return display_method(**kwargs)
-            self._drawers.show(**kwargs)
+            self._drawer.show(**kwargs)
 
         if not isinstance(self, Animation):
 
             from IPython.display import display
-            import ipywidgets as widgets
 
             try:
-                widget = self._drawers.get_ipywidget()
+                widget = self._drawer.get_ipywidget()
             except:
                 return _try_drawer()
 
@@ -1310,7 +1304,7 @@ class Plot(ShortCutable, Configurable, metaclass=PlotMeta):
                     self.call_shortcut(shortcut_key)
                     messages.value = ""
 
-                    self._update_FigureWidget(fig_widget)
+                    self._update_ipywidget(fig_widget)
 
             except Exception as e:
                 messages.value = f'<span style="color:darkred; font-weight: bold">{e}</span>'
@@ -1318,19 +1312,6 @@ class Plot(ShortCutable, Configurable, metaclass=PlotMeta):
         d.on_dom_event(partial(handle_dom_events))
 
         display(fig_widget, messages, h, styles, Output())
-
-    def _update_FigureWidget(self, fig_widget, plot = None):
-        """ Updates a figure widget so that it is in sync with this plot's data
-
-        Parameters
-        ----------
-        fig_widget: plotly.graph_objs.FigureWidget
-            The figure widget that we need to extend.
-        """
-        fig_widget.data = []
-        fig_widget.add_traces(self.data)
-        fig_widget.layout = self.layout
-        fig_widget.update(frames=self.frames)
 
     #-------------------------------------------
     #       PLOT MANIPULATION METHODS
@@ -1380,240 +1361,6 @@ class Plot(ShortCutable, Configurable, metaclass=PlotMeta):
         }[to]
 
         return PlotClass(plots=child_plots, **kwargs)
-
-    def group_legend(self, by=None, names=None, show_all=False, extra_updates=None, **kwargs):
-        """ Joins plot traces in groups in the legend
-
-        As the result of this method, plot traces end up with a legendgroup attribute.
-        You can use that for selecting traces in further processing of your plot.
-
-        This also provides the ability to toggle the whole group from the legend, which is nice.
-
-        Parameters
-        ---------
-        by: str or function, optional
-            it defines what are the criteria to group the traces.
-
-            If it's a string:
-                It is the name of the trace attribute. Remember that plotly allows you to
-                lookup for nested attributes using underscores. E.g: "line_color" gets {line: {color: THIS VALUE}}
-            If it's a function:
-                It will recieve each trace and needs to decide which group to put it in by returning the group value.
-                Note that the value will also be used as the group name if `names` is not provided, so you can save yourself
-                some code and directly return the group's name.
-            If not provided:
-                All traces will be put in the same group
-        names: array-like, dict or function, optional
-            it defines what the names of the generated groups will be.
-
-            If it's an array:
-                When a new group is found, the name will be taken from this array (order can be very arbitrary)
-            If it's a dict:
-                When a new group is found, the value of the group will be used as a key to get the name from this dictionary.
-                If the key is not found, the name will just be the value.
-                E.g.: If grouping by `line_color` and `blue` is found, the name will be `names.get('blue', 'blue')`
-            If it's a function:
-                It will recieve the group value and the trace and needs to return the name of the TRACE.
-                NOTE: If `show_all` is set to `True` all traces will appear in the legend, so it would be nice
-                to give them different names. Otherwise, you can just return the group's name.
-                If you provided a grouping function and `show_all` is False you don't need this, as you can return
-                directly the group name from there.
-            If not provided:
-                the values will be used as names.
-        show_all: boolean, optional
-            whether all the items of the group should be displayed in the legend.
-            If `False`, only one item per group will be displayed.
-            If `True`, all the items of the group will be displayed.
-        extra_updates: dict, optional
-            A dict stating extra updates that you want to do for each group.
-
-            E.g.: `{"blue": {"line_width": 4}}`
-
-            would also convert the lines with a group VALUE (not name) of "blue" to a width of 4.
-
-            This is just for convenience so that you can run other methods after this one.
-            Note that you can always do something like this by doing
-
-            ```
-            plot.update_traces(
-                selector={"line_width": "blue"}, # Selects the traces that you will update
-                line_width=4,
-            ) 
-            ```
-
-            If you use a function to return the group values, there is probably no point on using this
-            argument. Since you recieve the trace, you can run `trace.update(...)` inside your function.
-        **kwargs:
-            like extra_updates but they are passed to all groups without distinction
-        """
-        unique_values = []
-
-        # Normalize the "by" parameter to a function
-        if by is None:
-            if show_all:
-                name = names[0] if names is not None else "Group"
-                self.figure.update_traces(showlegend=True, legendgroup=name, name=name)
-                return self
-            else:
-                func = lambda trace: 0
-        if isinstance(by, str):
-            def func(trace):
-                try:
-                    return trace[by]
-                except Exception:
-                    return None
-        else:
-            func = by
-
-        # Normalize also the names parameter to a function
-        if names is None:
-            def get_name(val, trace):
-                return str(val) if not show_all else f'{val}: {trace.name}'
-        elif callable(names):
-            get_name = names
-        elif isinstance(names, dict):
-            def get_name(val, trace):
-                name = names.get(val, val)
-                return str(name) if not show_all else f'{name}: {trace.name}'
-        else:
-            def get_name(val, trace):
-                name = names[len(unique_values) - 1]
-                return str(name) if not show_all else f'{name}: {trace.name}'
-
-        # And finally normalize the extra updates
-        if extra_updates is None:
-            get_extra_updates = lambda *args, **kwargs: {}
-        elif isinstance(extra_updates, dict):
-            get_extra_updates = lambda val, trace: extra_updates.get(val, {})
-        elif callable(extra_updates):
-            get_extra_updates = extra_updates
-
-        # Build the function that will apply the change
-        def check_and_apply(trace):
-
-            val = func(trace)
-
-            if isinstance(val, np.ndarray):
-                val =  val.tolist()
-            if isinstance(val, list):
-                val = ", ".join([str(item) for item in val])
-
-            if val in unique_values:
-                showlegend = show_all
-            else:
-                unique_values.append(val)
-                showlegend = True
-
-            customdata = trace.customdata if trace.customdata is not None else [{}]
-
-            trace.update(
-                showlegend=showlegend,
-                legendgroup=str(val),
-                name=get_name(val, trace=trace),
-                customdata=[{**customdata[0], "name": trace.name}, *customdata[1:]],
-                **get_extra_updates(val, trace=trace),
-                **kwargs
-            )
-
-        # And finally apply all the changes
-        self.figure.for_each_trace(
-            lambda trace: check_and_apply(trace)
-        )
-
-        return self
-
-    def ungroup_legend(self):
-        """ Ungroups traces if a legend contains groups """
-        self.figure.for_each_trace(
-            lambda trace: trace.update(
-                legendgroup=None,
-                showlegend=True,
-                name=trace.customdata[0]["name"]
-            )
-        )
-
-        return self
-
-    def normalize(self, min_val=0, max_val=1, axis="y", **kwargs):
-        """ Normalizes traces to a given range along an axis
-
-        Parameters
-        -----------
-        min_val: float, optional
-            The lower bound of the range.
-        max_val: float, optional
-            The upper part of the range
-        axis: {"x", "y", "z"}, optional
-            The axis along which we want to normalize.
-        **kwargs:
-            keyword arguments that are passed directly to plotly's Figure `for_each_trace`
-            method. You can check its documentation. One important thing is that you can pass a
-            'selector', which will choose if the trace is updated or not. 
-        """
-        from .plotutils import normalize_trace
-
-        self.for_each_trace(partial(normalize_trace, min_val=min_val, max_val=max_val, axis=axis), **kwargs)
-
-        return self
-
-    def swap_axes(self, ax1='x', ax2='y', **kwargs):
-        """ Swaps two axes in the plot
-
-        Parameters
-        -----------
-        ax1, ax2: str, {'x', 'x*', 'y', 'y*', 'z', 'z*'}
-            The names of the axes that you want to swap. 
-        **kwargs:
-            keyword arguments that are passed directly to plotly's Figure `for_each_trace`
-            method. You can check its documentation. One important thing is that you can pass a
-            'selector', which will choose if the trace is updated or not. 
-        """
-        from .plotutils import swap_trace_axes
-        # Swap the traces
-        self.for_each_trace(partial(swap_trace_axes, ax1=ax1, ax2=ax2), **kwargs)
-
-        # Try to also swap the axes
-        try:
-            self.update_layout({
-                f'{ax1}axis': self.layout[f'{ax2}axis'].to_plotly_json(),
-                f'{ax2}axis': self.layout[f'{ax1}axis'].to_plotly_json(),
-            }, overwrite=True)
-        except:
-            pass
-
-        return self
-
-    def shift(self, shift, axis="y", **kwargs):
-        """ Shifts the traces of the plot by a given value in the given axis
-
-        Parameters
-        -----------
-        shift: float or array-like
-            If it's a float, it will be a solid shift (i.e. all points moved equally).
-            If it's an array, an element-wise sum will be performed
-        axis: {"x","y","z"}, optional
-            The axis along which we want to shift the traces.
-        **kwargs:
-            keyword arguments that are passed directly to plotly's Figure `for_each_trace`
-            method. You can check its documentation. One important thing is that you can pass a
-            'selector', which will choose if the trace is updated or not. 
-        """
-        from .plotutils import shift_trace
-
-        self.for_each_trace(partial(shift_trace, shift=shift, axis=axis), **kwargs)
-
-        return self
-
-    def v_line(self, x, yrange=None, **kwargs):
-        """ Draws a vertical line in the figure (NOT WORKING YET!) """
-        if yrange is None:
-            yrange = self.layout.yaxis.range
-            if yrange is None:
-                yrange = [0, 7000]
-
-        self.add_scatter(mode = "lines", x = [x, x], y = yrange, **kwargs)
-
-        return self
 
     def copy(self):
         """ Returns a copy of the plot
@@ -1671,10 +1418,6 @@ class Plot(ShortCutable, Configurable, metaclass=PlotMeta):
         if isinstance(path, str):
             path = Path(path)
 
-        if html or path.suffix == ".html":
-            self.figure.write_html(str(path.with_suffix(".html")))
-            return self
-
         #The following method actually modifies 'self', so there's no need to get the return
         self._get_pickleable()
 
@@ -1682,28 +1425,6 @@ class Plot(ShortCutable, Configurable, metaclass=PlotMeta):
             dill.dump(self, handle, protocol=dill.HIGHEST_PROTOCOL)
 
         return True
-
-    def html(self, path):
-        """ Just a shortcut for save( html = True )
-
-        Arguments
-        --------
-        path: str
-            The path to the file where you want to save the plot.
-        """
-        return self.save(path, html = True)
-
-    def to_chart_studio(self, *args, **kwargs):
-        """ Sends the plot to chart studio if it is possible
-
-        For it to work, the user should have their credentials correctly set up.
-
-        It is a shortcut for chart_studio.plotly.plot(self.figure, ...etc) so you can pass any extra arguments as if
-        you were using `py.plot`
-        """
-        import chart_studio.plotly as py
-
-        return py.plot(self.figure, *args, **kwargs)
 
 
 class EntryPoint:
@@ -1754,8 +1475,6 @@ def entry_point(name):
 #------------------------------------------------
 #       CLASSES TO SUPPORT COMPOSITE PLOTS
 #------------------------------------------------
-
-
 class MultiplePlot(Plot):
     """ General handler of a group of plots that need to be rendered together
 
@@ -1809,7 +1528,7 @@ class MultiplePlot(Plot):
         """ Returns all the attributes that its child_plots should have """
         return {
             'isChildPlot': True,
-            'shared_attr': lambda key: self.shared_attr(key),
+            '_get_shared_attr': lambda key: self.shared_attr(key),
             'share_attr': lambda key, val: self.set_shared_attr(key, val)
         }
 
@@ -1882,7 +1601,6 @@ class MultiplePlot(Plot):
                         leading_plot._SHOULD_SHARE_WITH_SIBLINGS = True
                         leading_plot.read_data(update_fig=False)
                         leading_plot._SHOULD_SHARE_WITH_SIBLINGS = False
-
                     self.set_data()
 
             else:
@@ -1944,11 +1662,7 @@ class MultiplePlot(Plot):
 
             If `True`, `plots` is added after them.
         """
-        # Maybe one of the plots is a plotly figure, normalize all to the plot class
-        plots = [Plot.from_plotly(plot) if isinstance(plot, go.Figure) else plot for plot in plots]
-
         for plot in plots:
-
             for key, val in self._attrs_for_child_plots.items():
                 setattr(plot, key, val)
 
@@ -1976,9 +1690,6 @@ class MultiplePlot(Plot):
         plot: sisl Plot or plotly Figure
             The plot to insert in the list
         """
-        if isinstance(plot, go.Figure):
-            plot = Plot.from_plotly(plot)
-
         self.child_plots.insert(index, plot)
 
     def shared_attr(self, key):
@@ -2015,35 +1726,6 @@ class MultiplePlot(Plot):
             the new value for the attribute
         """
         self.shared[key] = val
-
-        return self
-
-    def to_animation(self):
-        """ Converts the multiple plot into an animation by splitting its plots into frames """
-        self._isAnimation = True
-        self.clear().get_figure()
-        return self
-
-    def get_figure(self):
-        """ This method is responsible of building the figure from the child plots """
-        self.clear()
-
-        if getattr(self, "_isAnimation", False):
-            frames_layout = Animation._build_frames(self)
-            self.update_layout(**frames_layout)
-        elif getattr(self, "_is_subplots", False):
-            SubPlots._get_figure(self)
-
-        else:
-            data = []
-            for plot in self.child_plots:
-                data = [*data, *plot.data]
-
-            self.data = data
-
-        call_method_if_present(self, '_after_get_figure')
-
-        call_method_if_present(self, 'on_figure_change')
 
         return self
 
@@ -2131,206 +1813,13 @@ class Animation(MultiplePlot):
     def __init__(self, *args, frame_names=None, _plugins={}, **kwargs):
         if frame_names is not None:
             _plugins["_get_frame_names"] = frame_names if callable(frame_names) else lambda self, i: frame_names[i]
+        elif "_get_frame_names" not in _plugins:
+            _plugins["_get_frame_names"] = lambda self, i: f"Frame {i}"
 
         super().__init__(*args, **kwargs, _plugins=_plugins)
-
-    def _build_frames(self, ani_method):
-        """ Builds the frames of the plotly figure from the child plots' data
-
-        It actually sets the frames of the figure.
-
-        Returns
-        -----------
-        dict
-            keys and values that need to be added to the layout
-            in order for frames to work.
-        """
-        # Get the names for each frame
-        frame_names = []
-        for i, plot in enumerate(self.child_plots):
-            try:
-                frame_name = self._get_frame_names(i)
-            except Exception:
-                frame_name = f"Frame {i+1}"
-            frame_names.append(frame_name)
-
-        if ani_method is None:
-            same_traces = np.unique(
-                [len(plot.data) for plot in self.child_plots]
-            ).shape[0] == 1
-
-            ani_method = "animate" if same_traces else "update"
-
-        # Choose the method that we need to run in order to get the figure
-        if ani_method == "animate":
-            figure_builder = self._figure_animate_method
-        elif ani_method == "update":
-            figure_builder = self._figure_update_method
-
-        steps, updatemenus = figure_builder(frame_names)
-
-        frames_layout = {
-
-            "sliders": [
-                {
-                    "active": 0,
-                    "yanchor": "top",
-                    "xanchor": "left",
-                    "currentvalue": {
-                        "font": {"size": 20},
-                        #"prefix": "Bands file:",
-                        "visible": True,
-                        "xanchor": "right"
-                    },
-                    #"transition": {"duration": 300, "easing": "cubic-in-out"},
-                    "pad": {"b": 10, "t": 50},
-                    "len": 0.9,
-                    "x": 0.1,
-                    "y": 0,
-                    "steps": steps
-                }
-            ],
-
-            "updatemenus": updatemenus
-        }
-
-        return frames_layout
-
-    def _figure_update_method(self, frame_names):
-        """
-        In the update method, we give all the traces to data, and we are just going to toggle
-        their visibility depending on which 'frame' needs to be displayed.
-        """
-        # Add all the traces
-        for i, (frame_name, plot) in enumerate(zip(frame_names, self.child_plots)):
-
-            visible = i == 0
-
-            self.add_traces([{
-                **trace.to_plotly_json(),
-                'customdata': [{'frame': frame_name, "iFrame": i}],
-                'visible': visible
-            } for trace in plot.data])
-
-        # Generate the steps
-        steps = []
-        for i, frame_name in enumerate(frame_names):
-
-            steps.append({
-                "label": frame_name,
-                "method": "restyle",
-                "args": [{"visible": [trace.customdata[0]["iFrame"] == i for trace in self.data]}]
-            })
-
-        # WE SHOULD DEFINE PLAY AND PAUSE BUTTONS TO BE RENDERED IN JUPYTER'S NOTEBOOK HERE
-        # IT IS IMPOSSIBLE TO PASS CONDITIONS TO DECIDE WHAT TO DISPLAY USING PLOTLY JSON
-        self.animate_widgets = []
-
-        return steps, []
-
-    def _figure_animate_method(self, frame_names, frame_duration, redraw):
-        """
-        In the animate method, we explicitly define frames, And the transition from one to the other
-        will be animated
-        """
-        # Data will actually only be the first frame
-        self.data = self.child_plots[0].data
-
-        frames = []
-
-        maxN = np.max([len(plot.data) for plot in self.child_plots])
-        for frame_name, plot in zip(frame_names, self.child_plots):
-
-            data = plot.data
-            nTraces = len(data)
-            if nTraces < maxN:
-                nAddTraces = maxN - nTraces
-                data = [
-                    *data, *np.full(nAddTraces, {"type": "scatter", "x":  [0], "y": [0], "visible": False})]
-
-            frames = [
-                *frames, {'name': frame_name, 'data': data, "layout": plot.get_settings_group("layout")}]
-
-        self.frames = frames
-
-        steps = [
-            {"args": [
-            [frame["name"]],
-            {"frame": {"duration": int(frame_duration), "redraw": redraw},
-            "mode": "immediate",
-            "transition": {"duration": 300}}
-        ],
-            "label": frame["name"],
-            "method": "animate"} for frame in self.frames
-        ]
-
-        updatemenus = [
-
-            {'type': 'buttons',
-            'buttons': [
-                {
-                    'label': '▶',
-                    'method': 'animate',
-                    'args': [None, {"frame": {"duration": int(frame_duration), "redraw": True},
-                                    "fromcurrent": True, "transition": {"duration": 100,
-                                                                        "easing": "quadratic-in-out"}}],
-                },
-
-                {
-                    'label': '⏸',
-                    'method': 'animate',
-                    'args': [[None], {"frame": {"duration": 0}, "redraw": True,
-                                    'mode': 'immediate',
-                                    "transition": {"duration": 0}}],
-                }
-            ]}
-        ]
-
-        return steps, updatemenus
-
-    def merge_frames(self):
-        """ Merges all frames of an animation into one """
-        self._isAnimation = False
-        self.clear().get_figure()
-        return self
-
-    def zip(*animations):
-        """ Zips multiple animations together
-
-        This can work both as a static method or an object method:
-        It can be called from the Animation class. However, if called from an Animation, 
-        it will automatically include itself in the animations to zip.
-
-        YOU NEED TO MAKE SURE THAT ALL THE ANIMATIONS HAVE THE SAME NUMBER OF FRAMES
-
-        Parameters
-        -----------
-        *animations: sisl Animation
-            the animations that you want to zip together.
-            YOU NEED TO MAKE SURE THAT ALL THE ANIMATIONS HAVE THE SAME NUMBER OF FRAMES
-        """
-        frames = []
-        for (*old_frames,) in zip(*animations):
-
-            new_frame = MultiplePlot(plots=old_frames)
-
-            frames.append(new_frame)
-
-        return Animation(
-            plots=frames
-        )
-
-    def unzip(self):
-        """ Unzips the animation
-
-        In order for this method to make sense, the animation needs to be previously zipped.
-        This basically means that each frame needs to be a multiple plot and all frames are made
-        of the same number of plots.
-        """
-        # Basically we just need to get the plots for each frame and then transpose it
-        # so that we have the "frames for each plot"
-        new_animations = np.array([frame.child_plots for frame in self]).T
-        return [Animation(plots=plots) for plots in new_animations]
+    
+    def draw(self, drawer_info):
+        self._drawer.draw(drawer_info, self.child_plots, self._get_frame_names)
 
 
 class SubPlots(MultiplePlot):
@@ -2404,7 +1893,7 @@ class SubPlots(MultiplePlot):
         )
     )
 
-    def _get_figure(self, rows, cols, arrange, make_subplot_kwargs):
+    def draw(self, drawer_info, rows, cols, arrange, make_subplot_kwargs):
         """ Builds the subplots layout from the child plots' data """
         nplots = len(self.child_plots)
         if rows is None and cols is None:
@@ -2432,55 +1921,4 @@ class SubPlots(MultiplePlot):
         if cols * rows < nplots:
             warn(f"requested {nplots} on a {rows}x{cols} grid layout. {nplots - cols*rows} plots will be missing.")
 
-        # Check if all childplots have the same xaxis or yaxis titles.
-        axes_titles = defaultdict(list)
-        for child_plot in self.child_plots:
-            axes_titles["x"].append(child_plot.layout.xaxis.title.text)
-            axes_titles["y"].append(child_plot.layout.yaxis.title.text)
-
-        # If so, we will set the subplots figure x_title and/or y_title so that it looks cleaner.
-        # See how we remove the titles from the axis layout below when we allocate each plot.
-        axes_titles = {f"{key}_title": val[0] for key, val in axes_titles.items() if len(set(val)) == 1}
-
-        self.figure = make_subplots(**{
-            "rows": rows, "cols": cols,
-            **axes_titles,
-            **make_subplot_kwargs
-        })
-
-        # Start assigning each plot to a position of the layout
-        for (row, col), plot in zip(itertools.product(range(1, rows + 1), range(1, cols + 1)), self.child_plots):
-
-            ntraces = len(plot.data)
-
-            self.add_traces(plot.data, rows=[row]*ntraces, cols=[col]*ntraces)
-
-            for ax in "x", "y":
-                ax_layout = getattr(plot.layout, f"{ax}axis").to_plotly_json()
-
-                # If we have set a global title for this axis, just remove it from the plot
-                if axes_titles.get(f"{ax}_title"):
-                    ax_layout["title"] = None
-
-                update_axis = getattr(self, f"update_{ax}axes")
-
-                update_axis(ax_layout, row=row, col=col)
-
-        # Since we have directly copied the layouts of the child plots, there may be some references
-        # between axes that we need to fix. E.g.: if yaxis was set to follow xaxis in the second child plot,
-        # since the second child plot is put in (xaxes2, yaxes2) the reference will be now to the first child
-        # plot xaxis, not itself. This is best understood by printing the figure of a subplot :)
-        new_layouts = {}
-        for ax, layout in self.figure.layout.to_plotly_json().items():
-            if "axis" in ax:
-                ax_name, ax_num = ax.split("axis")
-
-                # Go over all possible problematic keys
-                for key in ["anchor", "scaleanchor"]:
-                    val = layout.get(key)
-                    if val in ["x", "y"]:
-                        layout[key] = f"{val}{ax_num}"
-
-                new_layouts[ax] = layout
-
-        self.update_layout(**new_layouts)
+        self._drawer.draw_subplots(drawer_info, rows, cols, self.child_plots)
