+++ sisl/viz/plots/__init__.py
@@ -2,9 +2,9 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at https://mozilla.org/MPL/2.0/.
 from .bands import BandsPlot
-from .fatbands import FatbandsPlot
-from .pdos import PdosPlot
-from .bond_length import BondLengthMap
-from .forces import ForcesPlot
-from .grid import GridPlot, WavefunctionPlot
-from .geometry import GeometryPlot
+# from .fatbands import FatbandsPlot
+# from .pdos import PdosPlot
+# from .bond_length import BondLengthMap
+# from .forces import ForcesPlot
+# from .grid import GridPlot, WavefunctionPlot
+# from .geometry import GeometryPlot
