+++ sisl/viz/plots/bands.py
@@ -6,7 +6,6 @@ from functools import partial
 import itertools
 
 import numpy as np
-import plotly.express as px
 
 import sisl
 from ..plot import Plot, entry_point
@@ -599,12 +598,17 @@ class BandsPlot(Plot):
         else:
             spin_moments = []
 
-        self._drawers.draw_bands(filtered_bands, self.spin_texture, spin_moments, self.spin.is_polarized, bands_color, spindown_color, bands_width, spin, add_band_trace_data)
+        return {
+            "draw_bands": [filtered_bands, self.spin_texture, spin_moments, self.spin.is_polarized, bands_color, spindown_color, bands_width, spin, add_band_trace_data],
+        }
+
+    def draw(self, drawer_info):
+        self._drawer.draw_bands(*drawer_info["draw_bands"])
 
         self._draw_gaps()
 
     def _after_get_figure(self, Erange, spin, spin_texture_colorscale):
-        self._drawers.after_get_figure(self, Erange, spin, spin_texture_colorscale)
+        self._drawer.after_get_figure(self, Erange, spin, spin_texture_colorscale)
 
     def _calculate_gaps(self):
         """
@@ -762,7 +766,7 @@ class BandsPlot(Plot):
             ks = [np.ravel(E.k)[0] for E in Es]
             Es = [np.ravel(E)[0] for E in Es]
 
-        self._drawers.draw_gap(ks, Es, color, name, **kwargs)
+        self._drawer.draw_gap(ks, Es, color, name, **kwargs)
 
         return self
 
@@ -786,6 +790,8 @@ class BandsPlot(Plot):
         Plot
             a new plot with the plotted information.
         """
+        import plotly.express as px
+
         two_bands = self.bands.sel(band=[band1, band2]).squeeze().values
 
         diff = two_bands[:, 1] - two_bands[:, 0]
@@ -819,6 +825,7 @@ class BandsPlot(Plot):
         Plot
             a new plot with the plotted information.
         """
+        import plotly.express as px
         b1, b2 = self.bands.sel(band=[band1, band2]).squeeze().values.T
         ks = self.bands.k.values
